moved review stars function into util funcs
